ai-core/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.17
openai==1.12.0
pymilvus==2.3.6
redis==5.0.1
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
httpx==0.26.0
python-dotenv==1.0.0
jieba==0.42.1
numpy==1.26.3
pandas==2.1.5
tiktoken==0.5.2
